# use optimized comparators for most common condition logic