Provides optimized logic processing for some of the most common logic:
* comparing a memory address to a constant
* comparing a memory address to its delta
* comparing a memory address to another memory address of the same size

Does not optimize indirect memory addresses (pointers), floats, or priors

When processing an arbitrary set of logic for 65536 frames, showed a ~33% decrease in time spent processing the achievements.

before (average 5.34us):
```
321.000000ms elapsed, 4.898071us average.
349.000000ms elapsed, 5.325317us average.
302.000000ms elapsed, 4.608154us average.
430.000000ms elapsed, 6.561279us average.
```
after (average 3.49us)
```
241.000000ms elapsed, 3.677368us average.
232.000000ms elapsed, 3.540039us average.
247.000000ms elapsed, 3.768921us average.
196.000000ms elapsed, 2.990723us average.
```